refactor: SupplierQueryClient
